backend/pyproject.toml
[tool.poetry]
name = "ping-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
ping3 = "^4.0.4"
django-cors-headers = "^4.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
